\name{mysecondpack-package}
\alias{mysecondpack-package}
\alias{mysecondpack}
\docType{package}
\title{
\packageTitle{mysecondpack}
}
\description{
\packageDescription{mysecondpack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mysecondpack}
\packageIndices{mysecondpack}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{mysecondpack}

Maintainer: \packageMaintainer{mysecondpack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
